PMA-7957 - Basic receive and display of push notification messages
- Cleaned up notification
- Fix API Subscribe event json decode

+++ pushnotificationsdk/src/main/java/com/everlytic/android/pushnotificationsdk/models/UnsubscribeEvent.kt
@@ -3,7 +3,7 @@ package com.everlytic.android.pushnotificationsdk.models
 import java.util.*
 
 internal data class UnsubscribeEvent(
-    val subscription_id: Int,
+    val subscription_id: Long,
     val device_id: String,
     val datetime: Date = Date()
 )
